apps/order-review/requirements.txt
fastapi
uvicorn
databricks-sdk
